+++ src/sparse.jl
@@ -8,7 +8,7 @@
 # This file is part of LazyAlgebra (https://github.com/emmt/LazyAlgebra.jl)
 # released under the MIT "Expat" license.
 #
-# Copyright (c) 2017-2018 Éric Thiébaut.
+# Copyright (c) 2017-2019 Éric Thiébaut.
 #
 
 """
@@ -32,34 +32,37 @@ are arrays whose dimensions are the `n` leading dimensions of `C` for input
 arrays whose dimensions are the remaining trailing dimensions of `C`.
 
 !!! note
-    For efficiency reasons, sparse operators are currently limited to dense
+    For efficiency reasons, sparse operators are currently limited to *flat*
     Julia arrays because they can be indexed linearly with no loss of
-    performances.  If `C`, `I` and/or `J` are not dense arrays, they will be
+    performances.  If `C`, `I` and/or `J` are not flat arrays, they will be
     automatically converted to regular arrays.
 
+See also [`isflatarray`](@ref).
 """
 struct SparseOperator{T,M,N,
-                      Tc<:DenseVector{T},
-                      Ti<:DenseVector{Int},
-                      Tj<:DenseVector{Int}} <: LinearMapping
+                      Tc<:AbstractVector{T},
+                      Ti<:AbstractVector{Int},
+                      Tj<:AbstractVector{Int}} <: LinearMapping
     outdims::NTuple{M,Int}
     inpdims::NTuple{N,Int}
     C::Tc
     I::Ti
     J::Tj
     samedims::Bool
+
+    # The inner constructor checks whether arguments are indeed flat arrays,
+    # it is not meant to be directly called.
     function SparseOperator{T,M,N,Tc,Ti,Tj}(
         outdims::NTuple{M,Int},
         inpdims::NTuple{N,Int},
         C::Tc,
         I::Ti,
         J::Tj
-    ) where {
-        T,M,N,
-        Tc<:DenseVector{T},
-        Ti<:DenseVector{Int},
-        Tj<:DenseVector{Int}
-    }
+    ) where {T,M,N,
+             Tc<:AbstractVector{T},
+             Ti<:AbstractVector{Int},
+             Tj<:AbstractVector{Int}}
+        @assert isflatarray(C, I, J)
         samedims = (M == N && outdims == inpdims)
         return new{T,M,N,Tc,Ti,Tj}(outdims, inpdims, C, I, J, samedims)
     end
@@ -67,14 +70,15 @@ end
 
 @callable SparseOperator
 
-function SparseOperator(outdims::NTuple{M,Int},
-                        inpdims::NTuple{N,Int},
-                        C::Tc,
-                        I::Ti,
-                        J::Tj) where {T,M,N,
-                                      Tc<:DenseVector{T},
-                                      Ti<:DenseVector{Int},
-                                      Tj<:DenseVector{Int}}
+# helper to call inner constructor
+function _sparseoperator(outdims::NTuple{M,Int},
+                         inpdims::NTuple{N,Int},
+                         C::Tc,
+                         I::Ti,
+                         J::Tj) where {T,M,N,
+                                       Tc<:AbstractVector{T},
+                                       Ti<:AbstractVector{Int},
+                                       Tj<:AbstractVector{Int}}
     return SparseOperator{T,M,N,Tc,Ti,Tj}(outdims, inpdims, C, I, J)
 end
 
@@ -83,22 +87,23 @@ function SparseOperator(outdims::NTuple{M,Integer},
                         C::AbstractVector{T},
                         I::AbstractVector{<:Integer},
                         J::AbstractVector{<:Integer}) where {T,M,N}
-    return SparseOperator(map(Int, outdims),
-                          map(Int, inpdims),
-                          densevector(T,   C),
-                          densevector(Int, I),
-                          densevector(Int, J))
+    return _sparseoperator(map(Int, outdims),
+                           map(Int, inpdims),
+                           flatvector(T,   C),
+                           flatvector(Int, I),
+                           flatvector(Int, J))
 end
 
-function SparseOperator(A::DenseArray{T,N}, n::Integer = 1) where {T,N}
+function SparseOperator(A::AbstractArray{T,N}, n::Integer = 1) where {T,N}
     1 ≤ n < N || throw(ArgumentError("bad number of of leading dimensions"))
+    has_standard_indexing(A) || throw(ArgumentError("array has non-standard indices"))
     dims = size(A)
     outdims = dims[1:n]
     inpdims = dims[n+1:end]
     nrows = prod(outdims)
     ncols = prod(inpdims)
     nz = 0
-    @inbounds for k in 1:length(A)
+    @inbounds for k in eachindex(A)
         if A[k] != zero(T)
             nz += 1
         end
@@ -106,10 +111,14 @@ function SparseOperator(A::DenseArray{T,N}, n::Integer = 1) where {T,N}
     C = Array{T}(undef, nz)
     I = Array{Int}(undef, nz)
     J = Array{Int}(undef, nz)
-    k = 0
-    l = 0
-    @inbounds for j in 1:ncols, i in 1:nrows
-        k += 1
+    i = j = l = 0
+    @inbounds for k in eachindex(A)
+        if i < nrows
+            i += 1
+        else
+            i = 1
+            j += 1
+        end
         if (a = A[k]) != zero(T)
             l += 1
             C[l] = a
@@ -129,7 +138,8 @@ input_size(S::SparseOperator) = S.inpdims
 output_size(S::SparseOperator) = S.outdims
 
 function is_same_mapping(A::SparseOperator{T,M,N,Tc,Ti,Tj},
-                         B::SparseOperator{T,M,N,Tc,Ti,Tj}) where {T,M,N,Tc,Ti,Tj}
+                         B::SparseOperator{T,M,N,Tc,Ti,Tj}) where {T,M,N,
+                                                                   Tc,Ti,Tj}
     return (is_same_mutable_object(coefs(A), coefs(B)) &&
             is_same_mutable_object(rows(A), rows(B)) &&
             is_same_mutable_object(cols(A), cols(B)) &&
@@ -142,9 +152,14 @@ EndomorphismType(S::SparseOperator) = (S.samedims ? Endomorphism : Morphism)
 _bad_input_dimensions() = throw(DimensionMismatch("bad input dimensions"))
 _bad_output_dimensions() = throw(DimensionMismatch("bad output dimensions"))
 
+_bad_input_indexing() =
+    throw(ArgumentError("input array has non-standard indices"))
+_bad_output_indexing() =
+    throw(ArgumentError("output array has non-standard indices"))
+
 function vcreate(::Type{Direct},
                  S::SparseOperator{Ts,M,N},
-                 x::DenseArray{Tx,N},
+                 x::AbstractArray{Tx,N},
                  scratch::Bool=false) where {Ts<:Real,Tx<:Real,M,N}
     # In-place operation is not possible so we simply ignore the scratch flag.
     size(x) == input_size(S) || _bad_input_dimensions()
@@ -153,7 +168,7 @@ end
 
 function vcreate(::Type{Adjoint},
                  S::SparseOperator{Ts,M,N},
-                 x::DenseArray{Tx,M},
+                 x::AbstractArray{Tx,M},
                  scratch::Bool=false) where {Ts<:Real,Tx<:Real,M,N}
     # In-place operation is not possible so we simply ignore the scratch flag.
     size(x) == output_size(S) || _bad_input_dimensions()
@@ -163,13 +178,15 @@ end
 function apply!(α::Real,
                 ::Type{Direct},
                 S::SparseOperator{Ts,M,N},
-                x::DenseArray{Tx,N},
+                x::AbstractArray{Tx,N},
                 scratch::Bool,
                 β::Real,
-                y::DenseArray{Ty,M}) where {Ts<:Number,Tx<:Number,
-                                            Ty<:Number,M,N}
-    size(x) == input_size(S) || _bad_input_dimensions()
+                y::AbstractArray{Ty,M}) where {Ts<:Number,Tx<:Number,
+                                               Ty<:Number,M,N}
+    size(x) == input_size(S)  || _bad_input_dimensions()
+    has_standard_indexing(x)  || _bad_input_indexing()
     size(y) == output_size(S) || _bad_output_dimensions()
+    has_standard_indexing(y)  || _bad_output_indexing()
     β == 1 || vscale!(y, β)
     α == 0 || _apply_sparse!(y,
                              convert_multiplier(α, promote_type(Ts, Tx), Ty),
@@ -180,25 +197,27 @@ end
 function apply!(α::Real,
                 ::Type{Adjoint},
                 S::SparseOperator{Ts,M,N},
-                x::DenseArray{Tx,M},
+                x::AbstractArray{Tx,M},
                 scratch::Bool,
                 β::Real,
-                y::DenseArray{Ty,N}) where {Ts<:Number,Tx<:Number,
-                                            Ty<:Number,M,N}
+                y::AbstractArray{Ty,N}) where {Ts<:Number,Tx<:Number,
+                                               Ty<:Number,M,N}
     size(x) == output_size(S) || _bad_input_dimensions()
-    size(y) == input_size(S) || _bad_output_dimensions()
+    has_standard_indexing(x)  || _bad_input_indexing()
+    size(y) == input_size(S)  || _bad_output_dimensions()
+    has_standard_indexing(y)  || _bad_output_indexing()
     β == 1 || vscale!(y, β)
     α == 0 || _apply_sparse!(y, convert_multiplier(α, promote_type(Ts, Tx), Ty),
                              coefs(S), cols(S), rows(S), x)
     return y
 end
 
-function _apply_sparse!(y::DenseArray{<:Real},
+function _apply_sparse!(y::AbstractArray{<:Real},
                         α::Real,
-                        C::DenseVector{<:Real},
-                        I::DenseVector{Int},
-                        J::DenseVector{Int},
-                        x::DenseArray{<:Real})
+                        C::AbstractVector{<:Real},
+                        I::AbstractVector{Int},
+                        J::AbstractVector{Int},
+                        x::AbstractArray{<:Real})
     # We already known that α is non-zero.
     length(I) == length(J) == length(C) ||
         error("corrupted sparse operator structure")
@@ -220,12 +239,12 @@ function _apply_sparse!(y::DenseArray{<:Real},
     end
 end
 
-function _apply_sparse!(y::DenseArray{<:Complex},
+function _apply_sparse!(y::AbstractArray{<:Complex},
                         α::Real,
-                        C::DenseVector{<:Complex},
-                        I::DenseVector{Int},
-                        J::DenseVector{Int},
-                        x::DenseArray{<:Real})
+                        C::AbstractVector{<:Complex},
+                        I::AbstractVector{Int},
+                        J::AbstractVector{Int},
+                        x::AbstractArray{<:Real})
     # We already known that α is non-zero.
     if α == 1
         for k in 1:length(C)
@@ -245,12 +264,12 @@ function _apply_sparse!(y::DenseArray{<:Complex},
     end
 end
 
-function _apply_sparse!(y::DenseArray{<:Complex},
+function _apply_sparse!(y::AbstractArray{<:Complex},
                         α::Real,
-                        C::DenseVector{<:Complex},
-                        I::DenseVector{Int},
-                        J::DenseVector{Int},
-                        x::DenseArray{<:Complex})
+                        C::AbstractVector{<:Complex},
+                        I::AbstractVector{Int},
+                        J::AbstractVector{Int},
+                        x::AbstractArray{<:Complex})
     # We already known that α is non-zero.
     if α == 1
         for k in 1:length(C)